requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
bcrypt
fpdf2
python-dotenv